# Optimize message serialization