… src/Orleans.Core/Messaging/IMessageSerializer.cs
… src/Orleans.Core/Messaging/PrefixingBufferWriter.cs
… src/Orleans.Serialization/Codecs/FieldHeaderCodec.cs
… src/Orleans.Serialization/Codecs/ReferenceCodec.cs
… src/Orleans.Serialization/Codecs/TagDelimitedFieldCodec.cs
… src/Orleans.Serialization/Session/ReferencedObjectCollection.cs
… src/Orleans.Serialization/Session/ReferencedTypeCollection.cs
… src/Orleans.Serialization/Utilities/VarIntWriterExtensions.cs
… src/Orleans.Serialization/WireProtocol/ExtendedWireType.cs
… src/Orleans.Serialization/WireProtocol/Field.cs
… src/Orleans.Serialization/WireProtocol/SchemaType.cs
… src/Orleans.Serialization/WireProtocol/WireType.cs
… test/Benchmarks/Serialization/Comparison/ArrayDeserializeBenchmark.cs
… test/Benchmarks/Serialization/Comparison/ClassSerializeBenchmark.cs
… test/Benchmarks/Serialization/ComplexTypeBenchmarks.cs
… test/Benchmarks/Serialization/FieldHeaderBenchmarks.cs
… test/Benchmarks/Serialization/Utilities/SingleSegmentBuffer.cs
… test/NonSilo.Tests/Serialization/MessageSerializerTests.cs