* Make `MessageBufferWriter` non-generic so that `Writer` methods actually get inlined (and don't have an additional generic parameter).
* Adjust `MessageSerializer` and `PrefixingBufferWriter` to make this work.
* Optimize field header and integer writing methods.

Benchmark results for message round-trip serialization are about 20% faster (__serialization itself is 40% / 1.66x faster__):

|               main |     Mean |    Error |   StdDev |   Gen0 | Allocated |
|--------------------- |---------:|---------:|---------:|-------:|----------:|
|    OrleansPipeWriter | 53.06 μs | 0.271 μs | 0.240 μs |      - |     480 B |
|     OrleansSerialize | 26.99 μs | 0.069 μs | 0.058 μs | 0.4272 |   17032 B |
|  OrleansBufferWriter | 24.44 μs | 0.402 μs | 0.376 μs |      - |         - |
| OrleansBufferWriter2 | 52.98 μs | 0.258 μs | 0.241 μs |      - |         - |
| OrleansMessageSerializerStructRoundTrip | 1.302 μs | 0.0245 μs | 0.0217 μs | 0.0095 |     336 B |
|  OrleansMessageSerializerClassRoundTrip | 1.994 μs | 0.0390 μs | 0.0494 μs | 0.0114 |     448 B |

|               PR |     Mean |    Error |   StdDev |   Gen0 | Allocated |
|--------------------- |---------:|---------:|---------:|-------:|----------:|
|    OrleansPipeWriter | 50.46 μs | 0.264 μs | 0.234 μs |      - |     480 B |
|     OrleansSerialize | 25.99 μs | 0.057 μs | 0.053 μs | 0.4272 |   17032 B |
|  OrleansBufferWriter | 22.09 μs | 0.116 μs | 0.103 μs |      - |         - |
| OrleansBufferWriter2 | 49.57 μs | 0.540 μs | 0.505 μs |      - |         - |
| OrleansMessageSerializerStructRoundTrip | 1.096 μs | 0.0048 μs | 0.0040 μs | 0.0095 |     336 B |
|  OrleansMessageSerializerClassRoundTrip | 1.637 μs | 0.0227 μs | 0.0201 μs | 0.0114 |     448 B |


###### Microsoft Reviewers: [Open in CodeFlow](https://portal.fabricbot.ms/api/codeflow?pullrequest=https://github.com/dotnet/orleans/pull/8185)